Project Checkpoint
Ensure cam id in the database function is treated as string
Add a function to copy database from local to prod
Now the dataabse have 3 more columns is_online is_flooded is_usable. But the script might not be able to handle these columns. Have to check for this later